Change file-name display to be consistent with other fields
All of the other columns in this table use <output> to store
output so we can add other elements (e.g. progress bars) to
and from the DOM w/o interfering with the text content. 

This commit changes the filename field to also use <output>
and simplifies the CSS so that all <output> tags use a fixed
width font, which will be particularly useful when we have
filenames with paths visible.