Select Files: disambiguate function names
Previously SelectFiles.onFilesChange and SelectFiles.props.onFilesChange
were not the same. Now the method which converts a DOM FileList to the
format we need is SelectFiles.processFileList and that method is the
only place which calls the onFilesChange handler passed down from
Bagger.

This commit also changes the file input handler to have a more specific
name