Avoid accessing file.size more than once
This translates into the browser making a syscall which is fairly
expensive with many files. This commit finishes the process of
ensuring that we only access file.size once to avoid hoping that the
browser properly memoizes it and preparing us to ensure that the UI is
driven off of our bag file records rather than accessing `File` in many
places.