Basic S3 upload support
This allows files to be uploaded but does not yet support things like resuming or fixity checks

Limitations of the AWS S3 SDK prevent use in a web worker without doing something like rewriting all of the window global accessors. Since uploading is no CPU bound it's much easier not to use a web worker for that portion of the job. The upload worker performance stats still need to be updated to handle this.