Test S3 credentials and provide basic UI feedback
We can’t do a listBuckets operation due to AWS limitations
but we can check the status of a bucket as soon as we have
all of the values.

This makes the test button work and updates the control 
state after various things like network errors, CORS, 
actual server responses, etc. have triggered a response.