Refactor uploader code
Bagger retains the methods which have bag-specific knowledge
(e.g. uploadPayloadFile() takes the local filename and adds
the “data/” prefix; uploadFile() prefixes with the bag name)
and the newly-renamed StorageManager class handles the
generic S3 settings for credentials, the key prefix if any,
etc. and will in the future handling any shimming we need
for use with other object storage systems or local storage
APIs.