UI improvements for upload/finalize controls
This changes the UI from command buttons from immediate 
actions to toggles. This means that you can check the
finalize toggle now and it will automatically complete the
bag as soon as hashing and uploading completes without
requiring you to sit around waiting for a large bag to 
complete.

This also means that we can get rid of the hiding since 
the future intent can be specified while other operations 
are in progress.