Generate tag manifests
These are small enough that we could use WebCrypto but that will require
dealing with a new API for encoding/decoding and producing a hex output
string. At the very least, this seems like it should wait until #46
since that could be a reasonable solution for both this need and hash
workers.

Closes #45